requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
Authlib==1.3.1
bcrypt==3.2.0
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
environs==11.0.0
exceptiongroup==1.2.0
fastapi==0.110.0
fastapi-cli==0.0.4
fastapi_cors==0.0.6
firebase-admin==6.5.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.2
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.5
grpcio-status==1.65.5
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
msgpack==1.0.8
orjson==3.10.5
packaging==24.1
passlib==1.7.4
proto-plus==1.24.0
protobuf==5.27.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
sendgrid==6.11.0
shellingham==1.5.4
sib-api-v3-sdk==7.6.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starkbank-ecdsa==2.2.0
starlette==0.36.3
typer==0.12.3
typing_extensions==4.10.0
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
